requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.6

# AI Provider - Google Gemini (FREE)
google-generativeai==0.3.2

# Alternative AI Provider - Groq (FREE, backup)
groq==0.4.1

# HTTP Client (for callbacks and testing)
requests==2.31.0
httpx==0.26.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
validators==0.22.0
email-validator==2.1.0

# Date/Time Utilities
python-dateutil==2.8.2

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
pylint==3.0.3
mypy==1.8.0



# Logging and Monitoring

python-json-logger==2.0.7



# Security

python-jose[cryptography]==3.3.0

passlib[bcrypt]==1.7.4



# Utilities

typing-extensions==4.9.0
